Deploy user NFT contract when starting dev server
