+++ packages/freshmint/commands/dev.ts
@@ -1,9 +1,12 @@
 import { Command } from 'commander';
 
+import { loadConfig } from '../config';
 import { runDevServer } from '../devServer';
 
 export default new Command('dev').description('start the Freshmint development server').action(dev);
 
 async function dev() {
-  await runDevServer();
+  const config = await loadConfig();
+
+  await runDevServer(config);
 }
