+++ packages/freshmint/devServer/index.ts
@@ -2,15 +2,16 @@ import chalk from 'chalk';
 
 import { startEmulator, EmulatorTransaction } from './startEmulator';
 import { startDevWallet } from './startDevWallet';
-import { deployContracts } from './deployContracts';
-
-export type ServerReadyHandler<T> = (value: T) => void;
+import { deploySupportingContracts } from './deploySupportingContracts';
+import { deployNFTContract } from '../deploy';
+import { FreshmintConfig } from '../config';
+import { FlowGateway } from '../flow';
 
 // TODO: clean up the child process handling logic
 //
 // Ensure that the emulator and dev wallet servers are never left running
 
-export async function runDevServer() {
+export async function runDevServer(config: FreshmintConfig) {
   const emulator = await startEmulator({
     onTransaction: (tx: EmulatorTransaction) => {
       if (tx.error) {
@@ -31,7 +32,7 @@ export async function runDevServer() {
     `${chalk.green('ready')} - started emulator on 0.0.0.0:${emulator.port}, url: http://localhost:${emulator.port}`,
   );
 
-  // Disable emulator output when deploying contracts to keep console clean
+  // Disable emulator output before starting dev wallet and deploying contracts to keep console clean
   emulator.showOutput(false);
 
   const devWallet = await startDevWallet();
@@ -43,8 +44,18 @@ export async function runDevServer() {
   );
 
   try {
-    await deployContracts();
-    console.log(`${chalk.green('ready')} - project contracts deployed`);
+    // Deploy the supporting contracts to the emulator
+    await deploySupportingContracts();
+
+    console.log(`${chalk.green('ready')} - deployed Freshmint contracts to the emulator`);
+
+    const network = 'emulator';
+    const flow = new FlowGateway(network, config.getContractAccount(network));
+
+    // Deploy the user's NFT contract to the emulator
+    await deployNFTContract(config, flow, network);
+
+    console.log(`${chalk.green('ready')} - deployed ${chalk.cyan(config.contract.name)} to the emulator`);
   } catch (error) {
     console.log(`${chalk.redBright('error - failed to deploy contracts:')}\n\n${chalk.red(error)}`);
